# doc/api.yaml
openapi: 3.1.0
info:
  title: VaikaParts Core Domain API
  description: |
    ## Overview

    The VaikaParts Core Domain API provides the foundational business logic and data management capabilities for the VaikaParts automotive parts marketplace platform. This API serves as the central domain service responsible for managing user profiles, auto parts demands, seller offers, contact exchange workflows, and notification systems.

    ## Architecture

    This API represents the core domain microservice within the VaikaParts distributed system architecture. It handles business logic and data persistence while integrating with external services through well-defined interfaces:

    - **Authentication Provider**: Supabase Auth via JWT token validation and webhook integration
    - **Payment Processing**: External Payment Service via webhook callbacks
    - **Storage**: S3-compatible object storage for profile photos and attachment URLs
    - **Messaging**: Event-driven notification system supporting multiple channels

    ## Domain Model

    The core domain encompasses the following primary entities:

    ### Users
    - **Researchers**: End users searching for automotive parts
    - **Sellers**: Vendors offering automotive parts for sale
    - **Managers**: Administrative users with platform oversight capabilities

    ### Business Entities
    - **Demands**: Researcher requests for specific automotive parts
...
    - **Parts**: Structured automotive part specifications
...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - supabaseAuth: []
...
paths:
...
  /v1/demands:
    get:
      operationId: getResearcherDemands
      tags: [Demands]
      summary: List researcher demands with optional status filtering
      description: |
        Returns a paginated list of demands belonging to the authenticated Researcher.
        Results may be narrowed by lifecycle status.

        **Authorization**: Researcher role required.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: status
          in: query
          required: false
          description: Filter by demand lifecycle status. Omit to return all statuses.
          schema:
            $ref: '#/components/schemas/PostStatus'
      responses:
        200:
          description: Demand list retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemandPageResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServer'
    post:
      operationId: createDemand
      tags: [Demands]
      summary: Create a new automotive parts demand
      description: |
        Creates a new demand on behalf of the authenticated Researcher. The request
        is submitted as `multipart/form-data`. Part fields are bound using Spring
        dot-notation (`part.name`, `part.carBrand`, etc.). Images are optional and
        bound under `part.images`; maximum 5 files per demand.

        **Authorization**: Researcher role required.
      parameters:
        - name: published
          in: query
          required: false
          description: apply or not default published status on the to be created demand.
          schema:
            type: boolean
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DemandCreateRequest'
            encoding:
              part.images:
                contentType: image/jpeg, image/png, image/webp
      responses:
        201:
          description: Demand created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Demand'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/demands/{demandId}:
    get:
      operationId: getDemandById
      tags: [Demands]
      summary: Retrieve a demand by its identifier
      description: |
        Returns the full detail view for a single demand. Access is restricted to
        the owning Researcher — requests for demands belonging to another principal
        return 404 to prevent information disclosure.

        **Authorization**: Researcher role required.
      parameters:
        - name: demandId
          in: path
          required: true
          description: Unique identifier of the demand.
          schema:
            type: string
      responses:
        200:
          description: Demand retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Demand'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/demands/{demandId}/status:
    patch:
      operationId: updateDemandStatus
      tags: [Demands]
      summary: Update the lifecycle status of a demand
      description: |
        Transitions the demand to a new lifecycle status. Only permitted
        status transitions are accepted; invalid transitions return 400.

        **Authorization**: Researcher role required.
      parameters:
        - name: demandId
          in: path
          required: true
          description: Unique identifier of the demand.
          schema:
            type: string
        - name: status
          in: query
          required: true
          description: Target lifecycle status to apply to the demand.
          schema:
            $ref: '#/components/schemas/PostStatus'
      responses:
        200:
          description: Demand status updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Demand'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/demands/{demandId}/offers:
    get:
      operationId: getOffersForDemand
      tags: [Demands, Offers]
      summary: Retrieve paginated offers submitted for a specific demand
      description: |
        Returns a paginated list of all offers submitted in response to the
        specified demand. Access is restricted to the demand owner.

        **Authorization**: Researcher role required.
      parameters:
        - name: demandId
          in: path
          required: true
          description: Unique identifier of the demand.
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        200:
          description: Offer list retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferPageResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/demands/all:
    get:
      operationId: get all demands
      tags: [Demands]
      summary: List demands with pagination
      description: |
        Returns a paginated list of demands belonging to the authenticated Seller.
        Results may be narrowed by lifecycle status.

        **Authorization**: Seller role required.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        200:
          description: Demand list retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemandPageResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServer'
...
  /v1/offers/demand/{demandId}:
    get:
      operationId: getOffersByDemandId
      tags: [Offers]
      summary: Retrieve paginated offers associated with a specific demand
      description: |
        Returns a paginated list of offers submitted in response to the specified
        demand, as visible to the authenticated seller.

        **Authorization**: Seller role required.
      parameters:
        - name: demandId
          in: path
          required: true
          description: Unique identifier of the demand.
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        200:
          description: Offer list retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferPageResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/notifications/demand-published:
    get:
      operationId: fetchNotifications
      tags: [Notifications]
      summary: Retrieve user notifications
      description: Returns a paginated list of notifications for the current authenticated user.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        200:
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPageResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/notifications/demand-published/mark-as-read/{notificationId}:
    patch:
      operationId: markAsRead
      tags: [Notifications]
      summary: Mark notification as read
      description: Updates the read status of a specific notification.
      parameters:
        - name: notificationId
          in: path
          required: true
          description: Unique identifier of the notification
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        202:
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/notifications/demand-published/{notificationId}:
    get:
      operationId: getNotification
      tags: [Notifications]
      summary: Get notification with given id
      description: Allows the authenticated recipient to fetch the notification with the given id.
      parameters:
        - name: notificationId
          in: path
          required: true
          description: Unique identifier of the notification
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        202:
          description: Fetched notification with the corresponding given id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServer'
components:
  securitySchemes:
    supabaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase-issued JWT token for user authentication
    supabaseWebhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: HMAC signature for Supabase webhook verification
    paymentWebhookSignature:
      type: apiKey
      in: header
      name: X-Payment-Signature
      description: HMAC signature for Payment Service webhook verification
  parameters:
    PageParam:
      name: page
      in: query
      description: Zero-indexed page number for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  schemas:
...
    UUID:
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      description: RFC 4122 compliant UUID
      example: "e6b5cc3e-eaaa-4cb3-912c-e059e983bb78"
    DateTime:
      type: string
      format: date-time
      description: ISO 8601 formatted date-time string
      example: "2025-12-06T11:23:45Z"
...
    Location:
      type: object
      required: [city, region, address]
      description: Geographical location information
      properties:
        city:
          type: string
          description: Location city
          enum: [ANTANANARIVO]
        region:
          type: string
          description: Location region
          enum: [ANALAMANGA]
        address:
          type: string
          description: Detailed street address
...
    NotificationType:
      type: string
      enum: [DEMAND_PUBLISHED, OFFER_PUBLISHED, OFFER_ACCEPTED, OFFER_REJECTED, DEMAND_CANCELED, SYSTEM_ANNOUNCEMENT]
      description: Notification event type
...
    Researcher:
      allOf:
        - $ref: '#/components/schemas/BaseUser'
        - type: object
          required: [location]
          description: Researcher user profile
          properties:
            location:
              $ref: '#/components/schemas/Location'
    Seller:
      allOf:
        - $ref: '#/components/schemas/BaseUser'
        - type: object
          required: [location]
          description: Seller user profile
          properties:
            garage_name:
              type: string
              description: Business or garage name
            location:
              $ref: '#/components/schemas/Location'
            lat_lon:
              $ref: '#/components/schemas/LatLon'
    Manager:
      allOf:
        - $ref: '#/components/schemas/BaseUser'
        - type: object
          required: [role]
          description: Manager user profile with administrative permissions
          properties:
            role:
              $ref: '#/components/schemas/ManagerRole'
    PartCategory:
      type: string
      description: Automotive part category classification.
      enum:
        - ENGINE_PART
        - TURBO_COMPRESSOR
        - COOLING_SYSTEM
        - FUEL_SYSTEM
        - AIR_FILTER
        - OIL_FILTER
        - EXHAUST_PIPE
        - CATALYTIC_CONVERTER
        - CLUTCH_GEARBOX
        - TRANSMISSION
        - DRIVESHAFT
        - DIFFERENTIAL
        - STARTER
        - ALTERNATOR
        - BATTERY
        - IGNITION_SYSTEM
        - SPARK_PLUGS
        - WIRING_HARNESS
        - FUSES_RELAYS
        - SENSORS_AND_PROBES
        - ECU_COMPUTER
        - DIRECTION_SUSPENSION
        - SHOCK_ABSORBERS
        - SPRINGS
        - CONTROL_ARMS
        - BALL_JOINTS
        - STEERING_RACK
        - POWER_STEERING_PUMP
        - BRAKING
        - BRAKE_PADS
        - BRAKE_DISCS
        - BRAKE_DRUMS
        - BRAKE_CALIPERS
        - MASTER_CYLINDER
        - ABS_SYSTEM
        - TIRE
        - WHEELS_RIMS
        - WHEEL_BEARINGS
        - HEADLIGHTS
        - TAIL_LIGHTS
        - FOG_LIGHTS
        - INDICATORS
        - INTERIOR_LIGHTS
        - BODY
        - BUMPERS
        - DOORS
        - FENDERS
        - HOOD_BONNET
        - TRUNK_TAILGATE
        - MIRRORS
        - WINDSHIELD_WINDOWS
        - WIPERS
        - GRILLE
        - PASSENGER_COMPARTMENT
        - SEATS
        - DASHBOARD
        - STEERING_WHEEL
        - PEDALS
        - CARPETS_MATS
        - AC_HEATER
        - RADIATOR
        - THERMOSTAT
        - COOLING_FAN
        - AC_COMPRESSOR
        - ENGINE_OIL
        - BRAKE_FLUID
        - COOLANT
        - TRANSMISSION_FLUID
        - DISTRIBUTION
        - TIMING_BELT
        - TIMING_CHAIN
        - ACCESSORIES
        - UNIVERSAL_PARTS
        - OTHER
      example: FOG_LIGHTS
    PostStatus:
      type: string
      description: Lifecycle status of an exchange.
      enum:
        - DRAFT
        - PUBLISHED
        - SUSPENDED
        - PENDING
        - CANCELED
      example: PUBLISHED
    Part:
      type: object
      description: Automotive part being sought by the Researcher.
      required: [name, car_brand, car_model, car_year, part_category]
      properties:
        name:
          type: string
          description: Descriptive name of the part.
          example: Headlight
        car_brand:
          type: string
          description: Vehicle manufacturer.
          example: Toyota
        car_model:
          type: string
          description: Vehicle model designation.
          example: Corolla
        car_year:
          type: integer
          description: Four-digit model year of the vehicle.
          minimum: 1900
          maximum: 2100
          example: 2015
        part_category:
          $ref: '#/components/schemas/PartCategory'
        image_urls:
          type: array
          description: >
            Publicly accessible URLs of reference photographs uploaded by the Researcher. Populated server-side after images are stored.

          items:
            type: string
            format: uri
          maxItems: 5
          readOnly: true
    Demand:
      type: object
      description: A Researcher's request for a specific automotive part.
      required: [id, resource_type, researcher, part, status, created_at]
      properties:
        resource_type:
          type: string
          enum: [DEMAND]
          readOnly: true
        id:
          $ref: '#/components/schemas/UUID'
        description:
          type: string
          description: Additional context or specifications provided by the Researcher.
          example: Looking for headlight
        attached_photos_urls:
          type: array
          description: URLs of photographs attached to the demand.
          items:
            type: string
            format: uri
          maxItems: 5
          readOnly: true
        researcher:
          $ref: '#/components/schemas/Researcher'
        part:
          $ref: '#/components/schemas/Part'
        status:
          $ref: '#/components/schemas/PostStatus'
        created_at:
          $ref: '#/components/schemas/DateTime'
        updated_at:
          $ref: '#/components/schemas/DateTime'
        canceled_at:
          $ref: '#/components/schemas/DateTime'
        suspended_at:
          $ref: '#/components/schemas/DateTime'
    DemandCreateRequest:
      type: object
      description: >
        Payload for creating a new demand, submitted as multipart/form-data. Part fields use Spring dot-notation binding (e.g. `part.name`, `part.images`).

      required: [description, part]
      properties:
        description:
          type: string
          description: Additional context or specifications for the sought part.
          maxLength: 1000
          example: Looking for headlight
        part:
          $ref: '#/components/schemas/PartCreateRequest'
    PartCreateRequest:
      type: object
      description: >
        Part details submitted as part of the demand creation multipart request. Fields are bound using Spring dot-notation: `part.name`, `part.car_brand`, etc.

      required: [name, car_brand, car_model, car_year, part_category]
      properties:
        name:
          type: string
          description: Descriptive name of the part.
          example: Headlight
        car_brand:
          type: string
          description: Vehicle manufacturer.
          example: Toyota
        car_model:
          type: string
          description: Vehicle model designation.
          example: Corolla
        car_year:
          type: integer
          description: Four-digit model year of the vehicle.
          minimum: 1900
          maximum: 2100
          example: 2015
        part_category:
          $ref: '#/components/schemas/PartCategory'
        images:
          type: array
          description: >
            Reference photographs of the sought part. Bound under the multipart field name `part.images`. Optional; maximum 5 files.

          items:
            type: string
            format: binary
          maxItems: 5
    Offer:
      type: object
      required: [id, resource_type, seller_id, parts_info, demand, status, created_at]
      description: Seller response to a Researcher demand.
      properties:
        resource_type:
          type: string
          enum: [OFFER]
          readOnly: true
        id:
          $ref: '#/components/schemas/UUID'
        description:
          type: string
          description: Part condition details and additional information.
          maxLength: 1000
        attached_photos_urls:
          type: array
          items:
            type: string
            format: uri
          description: Photographs of the actual part.
          maxItems: 5
          readOnly: true
        seller_id:
          $ref: '#/components/schemas/UUID'
        parts_info:
          $ref: '#/components/schemas/PartInfo'
        demand:
          $ref: '#/components/schemas/Demand'
        status:
          $ref: '#/components/schemas/PostStatus'
        created_at:
          $ref: '#/components/schemas/DateTime'
        updated_at:
          $ref: '#/components/schemas/DateTime'
        canceled_at:
          $ref: '#/components/schemas/DateTime'
        suspended_at:
          $ref: '#/components/schemas/DateTime'
...
    Notification:
      type: object
      required: [id, recipient, message, notification_type, created_at, read]
      description: User notification record
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        recipient:
          oneOf:
            - $ref: '#/components/schemas/Researcher'
            - $ref: '#/components/schemas/Seller'
            - $ref: '#/components/schemas/Manager'
          discriminator:
            propertyName: user_type
            mapping:
              RESEARCHER: '#/components/schemas/Researcher'
              SELLER: '#/components/schemas/Seller'
              MANAGER: '#/components/schemas/Manager'
        notification_requested_id:
          type: string
          description: Notification event id
        message:
          type: string
          description: Notification body text
        resource:
          nullable: true
          oneOf:
            - $ref: '#/components/schemas/Demand'
            - $ref: '#/components/schemas/Offer'
          discriminator:
            propertyName: resource_type
            mapping:
              DEMAND: '#/components/schemas/Demand'
              OFFER: '#/components/schemas/Offer'
        notification_type:
          $ref: '#/components/schemas/NotificationType'
        read:
          type: boolean
          default: false
        created_at:
          $ref: '#/components/schemas/DateTime'
        click_action:
          example: {"action": "VIEW_DEMAND", "demandId": "demand-ctrl-789"}
        read_at:
          $ref: '#/components/schemas/DateTime'
    PageResponse:
      type: object
      required: [content, pageable, total_elements, total_pages, last, first, size, number, number_of_elements, empty]
      description: >
        Paginated response wrapper matching the JSON serialization of org.springframework.data.domain.Page. The pagination metadata is nested under the `pageable` object, consistent with Spring Data defaults.

      properties:
        content:
          type: array
          description: Page content — the list of elements for the current page.
          items: {}
        pageable:
          type: object
          description: Pagination state used to produce this page.
          properties:
            page_number:
              type: integer
              description: Zero-based current page index.
            page_size:
              type: integer
              description: Requested number of items per page.
            offset:
              type: integer
              format: int64
              description: Offset of the first element of this page relative to the full result set.
            sort:
              $ref: '#/components/schemas/SortDescriptor'
            paged:
              type: boolean
              description: Whether pagination is applied.
            unpaged:
              type: boolean
              description: Whether pagination is disabled.
        total_elements:
          type: integer
          format: int64
          description: Total number of elements across all pages.
        total_pages:
          type: integer
          description: Total number of pages available.
        last:
          type: boolean
          description: Whether this is the last page.
        first:
          type: boolean
          description: Whether this is the first page.
        size:
          type: integer
          description: Effective page size (number of elements requested per page).
        number:
          type: integer
          description: Zero-based index of the current page.
        number_of_elements:
          type: integer
          description: Number of elements on the current page.
        empty:
          type: boolean
          description: Whether the current page has no content.
        sort:
          $ref: '#/components/schemas/SortDescriptor'
    SortDescriptor:
      type: object
      description: Sort state descriptor returned by Spring Data.
      properties:
        empty:
          type: boolean
          description: Whether no sort orders are applied.
        sorted:
          type: boolean
          description: Whether at least one sort order is applied.
        unsorted:
          type: boolean
          description: Whether no sort orders are applied (inverse of sorted).
...
    DemandPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Demand'
    OfferPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Offer'
    NotificationPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Notification'
    ErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      description: Standard error response format
      properties:
        timestamp:
          $ref: '#/components/schemas/DateTime'
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: HTTP status reason phrase
        message:
          type: string
          description: Human-readable error description
        path:
          type: string
          description: The relative URI of the request that caused the error.
          example: "/api/v1/resource-path"
        errorCode:
          type: string
          description: Application-specific error code
          enum:
            - RESOURCE_NOT_FOUND
            - ILLEGAL_STATE
            - SECURITY_EXCEPTION
            - PAYLOAD_NOT_AS_EXPECTED
            - EMAIL_HEALTH_CHECK_FAILED
            - MISSING_REQUIRED_PART
            - DUPLICATE_EMAIL
            - DUPLICATE_SUPABASE_ID
            - DIRECTORY_UPLOAD_FAILED
            - AUTHORIZATION_DENIED
            - ENTITY_NOT_FOUND
            - MISSING_REQUEST_HEADER
            - MISSING_AUTH_CREDS
            - USER_NOT_FOUND
            - BAD_FORM
            - MISSING_AUTHORIZATION
            - INTERNAL_SERVER_ERROR
            - MISSING_REQUIRED_PARAMETER
            - INVALID_PARAMETER
            - MALFORMED_JSON
            - METHOD_NOT_ALLOWED
        details:
          type: object
          description: Additional error context
  responses:
    InternalServer:
      description: Internal Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request parameters or payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions for requested operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...